+++ js/packages/ui/src/contexts/idle/index.ts
@@ -1,12 +1,10 @@
 "use client";
 
 // Idle timeout context and provider
-export type { IdleTimeoutContextType } from "./IdleTimeoutContext";
-export {
-  IdleTimeoutProvider,
-  useIdleTimeout,
-  useIdleTimeoutSafe,
-} from "./IdleTimeoutContext";
+export { IdleTimeoutProvider, useIdleTimeout } from "./IdleTimeoutContext";
+// Types and safe hook (no dependencies on other idle modules)
+export type { IdleTimeoutContextType } from "./types";
+export { useIdleTimeoutSafe } from "./types";
 
 // Idle detection hook (typically used internally by IdleTimeoutProvider)
 export { useIdleDetection } from "./useIdleDetection";
